+++ src/db/migration/5_add_quantity_column_into_is_composed_by_table.sql
@@ -1,1 +1,1 @@
-alter table is_composed_by add column quantity integer;
+alter table is_composed_by add column quantity float;
